requirements/production.txt
# Try not to put anything here. Avoid dependencies in
# production that aren't in development.
-r base.txt

# WSGI Handler
# ------------------------------------------------
gevent==1.1.1
gunicorn==19.6.0

# Raven is the Sentry client
# --------------------------
raven

# Email backends for Mailgun, Postmark, SendGrid and more
# -------------------------------------------------------
django-anymail==0.4.2

